more crazy stuff (unfinished planetary)

+++ src/client/game/weapons/projectile/WeaponBurst.h
@@ -4,12 +4,12 @@
 
 #pragma once
 
-#include "ProjectileWeapon.h"
+#include "client/game/weapons/projectile/base/ProjectileWeapon.h"
 
 class WeaponBurst : public ProjectileWeapon
 {
 private:
-	double m_RecoilForce, m_BaseRecoilForce;
+	float m_RecoilForce, m_BaseRecoilForce;
 	double m_Damage;
 
 	unsigned long long m_BurstTick;
