+++ src/client/game/weapons/projectile/WeaponShotgun.h
@@ -4,13 +4,13 @@
 
 #pragma once
 
-#include "ProjectileWeapon.h"
+#include "client/game/weapons/projectile/base/ProjectileWeapon.h"
 
 class WeaponShotgun : public ProjectileWeapon
 {
 private:
 	int m_PelletCount;
-	double m_RecoilForce, m_BaseRecoilForce;
+	float m_RecoilForce, m_BaseRecoilForce;
 	double m_Damage;
 
 public:
